PR comment fixes
